requirements.txt
certifi==2020.11.8
chardet==3.0.4
docopt==0.6.2
idna==2.10
numpy==1.19.2
pandas==1.1.3
pyaml==20.4.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.25.0
ruamel.yaml==0.16.12
six==1.15.0
urllib3==1.26.2
